refactor: use one function
