+++ packages/twenty-front/src/modules/object-record/record-field/meta-types/hooks/useFormattedJsonFieldValue.ts
@@ -1,8 +1,7 @@
 import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSingular';
 import { FieldContext } from '@/object-record/record-field/contexts/FieldContext';
+import { orderWorkflowRunOutput } from '@/object-record/record-field/meta-types/utils/orderWorkflowRunOutput';
 import { FieldJsonValue } from '@/object-record/record-field/types/FieldMetadata';
-import { WorkflowRunOutput } from '@/workflow/types/Workflow';
-import { workflowRunOutputSchema } from '@/workflow/validation-schemas/workflowSchema';
 import { useContext } from 'react';
 import { isDefined } from 'twenty-shared/utils';
 
@@ -19,26 +18,7 @@ export const useFormattedJsonFieldValue = ({
     fieldDefinition.metadata.fieldName === 'output' &&
     isDefined(fieldValue)
   ) {
-    const parsedValue = workflowRunOutputSchema.safeParse(fieldValue);
-    if (!parsedValue.success) {
-      return null;
-    }
-
-    const orderedWorkflowRunOutput: WorkflowRunOutput = {
-      ...(isDefined(parsedValue.data.error)
-        ? {
-            error: parsedValue.data.error,
-          }
-        : {}),
-      ...(isDefined(parsedValue.data.stepsOutput)
-        ? {
-            stepsOutput: parsedValue.data.stepsOutput,
-          }
-        : {}),
-      flow: parsedValue.data.flow,
-    };
-
-    return orderedWorkflowRunOutput as FieldJsonValue;
+    return orderWorkflowRunOutput(fieldValue) as FieldJsonValue;
   }
 
   return fieldValue;
