+++ packages/twenty-front/src/modules/object-record/record-field/meta-types/utils/orderWorkflowRunOutput.ts
@@ -1,0 +1,27 @@
+import { WorkflowRunOutput } from '@/workflow/types/Workflow';
+import { workflowRunOutputSchema } from '@/workflow/validation-schemas/workflowSchema';
+import { isDefined } from 'twenty-shared/utils';
+import { JsonValue } from 'type-fest';
+
+export const orderWorkflowRunOutput = (value: JsonValue) => {
+  const parsedValue = workflowRunOutputSchema.safeParse(value);
+  if (!parsedValue.success) {
+    return null;
+  }
+
+  const orderedWorkflowRunOutput: WorkflowRunOutput = {
+    ...(isDefined(parsedValue.data.error)
+      ? {
+          error: parsedValue.data.error,
+        }
+      : {}),
+    ...(isDefined(parsedValue.data.stepsOutput)
+      ? {
+          stepsOutput: parsedValue.data.stepsOutput,
+        }
+      : {}),
+    flow: parsedValue.data.flow,
+  };
+
+  return orderedWorkflowRunOutput;
+};
